requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
drf-yasg>=1.21
django-filter>=24.2
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=22.0
whitenoise>=6.6
